+++ vpx_baseline_vary_resolution.py
@@ -6,6 +6,7 @@ import log_parser
 import numpy as np
 from utils import *
 import shutil
+from time import perf_counter
 
 
 parser = argparse.ArgumentParser(description='Compression Factor Variation.')
@@ -33,8 +34,8 @@ parser.add_argument('--num-runs', type=int,
 parser.add_argument('--root-dir', type=str,
                     help='name of default video directory', 
                     default="sundar_pichai.mp4")
-parser.add_argument('--person', type=str,
-                    help='person who will be resized', default='xiran')
+parser.add_argument('--people', type=str, nargs='+',
+                    help='person who will be resized', default=['xiran'])
 parser.add_argument('--save-prefix', type=str,
                     help='prefix to save logs and files in', 
                     required=True)
@@ -44,6 +45,8 @@ parser.add_argument('--executable-dir', type=str,
 parser.add_argument('--csv-name', type=str,
                     help='file to save final data in', 
                     default="data/frame_rate_data")
+parser.add_argument('--video-num-range', type=int, nargs=2,
+                    help='video start and end range', default=[0, 4])
 
 args = parser.parse_args()
 
@@ -60,34 +63,50 @@ def run_experiments():
     params['duration'] = args.duration
     params['runs'] = 1
     params['enable_prediction'] = False
-    person = args.person
+    params['socket_path'] = 'baseline.sock'
+    vid_start, vid_end = args.video_num_range
+    assert(vid_end >= vid_start)
     
-    for resolution in args.resolutions:
-        video_dir = os.path.join(args.root_dir, person, "test")
-        
-        for video_name in os.listdir(video_dir):
-            video_file = os.path.join(video_dir, video_name)
-            params['save_dir'] = f'{save_prefix}_vpx/{person}/resolution{resolution}/' + \
-                    f'{os.path.basename(video_name)}'
-            params['video_file'] = f'{params["save_dir"]}/{video_name}'
+    for person in args.people:
+        for resolution in args.resolutions:
+            video_dir = os.path.join(args.root_dir, person, "test")
+            
+            for i, video_name in enumerate(os.listdir(video_dir)):
+                if i not in range(vid_start, vid_end + 1):
+                    continue
+                
+                video_file = os.path.join(video_dir, video_name)
+                params['save_dir'] = f'{save_prefix}_vpx/{person}/resolution{resolution}/' + \
+                        f'{os.path.basename(video_name)}'
+                params['video_file'] = f'{params["save_dir"]}/{video_name}'
 
-            shutil.rmtree(params['save_dir'], ignore_errors=True)
-            os.makedirs(params['save_dir'])
-
-            width, height = resolution.split("x")
-            ffmpeg_cmd = f'ffmpeg -hide_banner -loglevel error -y -i {video_file} ' + \
-                    f'-vf scale={width}:{height} {params["video_file"]}'
-            print(ffmpeg_cmd)
-            os.system(ffmpeg_cmd)
-
-            for quantizer in args.quantizer_list:
-                params['save_dir'] = f'{save_prefix}_vpx/{person}/resolution{resolution}/quantizer{quantizer}/' + \
-                    f'{os.path.basename(video_name)}'
                 shutil.rmtree(params['save_dir'], ignore_errors=True)
                 os.makedirs(params['save_dir'])
-                
-                params['quantizer'] = quantizer
-                run_single_experiment(params)
+
+                start = perf_counter()
+                width, height = resolution.split("x")
+                ffmpeg_cmd = f'ffmpeg -hide_banner -loglevel error -y -i {video_file} ' + \
+                        f'-vf scale={width}:{height} {params["video_file"]}'
+                print(ffmpeg_cmd)
+                os.system(ffmpeg_cmd)
+                end = perf_counter()
+                print("ffmpeg command took", end - start)
+
+                for quantizer in args.quantizer_list:
+                    params['save_dir'] = f'{save_prefix}_vpx/{person}/resolution{resolution}/quantizer{quantizer}/' + \
+                        f'{os.path.basename(video_name)}'
+                    start = perf_counter()
+                    shutil.rmtree(params['save_dir'], ignore_errors=True)
+                    os.makedirs(params['save_dir'])
+                    end = perf_counter()
+                    print("rm command took", end - start)
+                    
+                    params['quantizer'] = quantizer
+                    start = perf_counter()
+                    print(f'Run {video_name} for person {person} resolution {resolution} quantizer {quantizer}')
+                    run_single_experiment(params)
+                    end = perf_counter()
+                    print("single experiment took", end - start)
 
 
 """ gets bitrate info from pcap file 
@@ -96,55 +115,63 @@ def run_experiments():
 def aggregate_data():
     first = True
     save_prefix = args.save_prefix
-    person = args.person
+    vid_start, vid_end = args.video_num_range
+    assert(vid_end >= vid_start)
     
     for resolution in args.resolutions:
         for quantizer in args.quantizer_list:
             combined_df = pd.DataFrame()
 
-            video_dir = os.path.join(args.root_dir, person, "test")
-            
-            for video_name in os.listdir(video_dir):
-                print(f'Run {video_name} for person {person} resolution {resolution}')
-                save_dir = f'{save_prefix}_vpx/{person}/resolution{resolution}/quantizer{quantizer}/' + \
-                        f'{os.path.basename(video_name)}/run0/'
-                dump_file = f'{save_dir}/sender.log'
-                saved_video_file = f'{save_dir}/received.mp4'
-                print(save_dir)
+            for person in args.people:
+                video_dir = os.path.join(args.root_dir, person, "test")
 
-                stats = log_parser.gather_trace_statistics(dump_file, args.window)
-                num_windows = len(stats['bitrates']['video'])
-                streams = list(stats['bitrates'].keys())
-                stats['bitrates']['time'] = np.arange(1, num_windows + 1)
-                window = stats['window']
-                
-                df = pd.DataFrame.from_dict(stats['bitrates'])
-                for s in streams:
-                    df[s] = (df[s] / 1000.0 / window).round(2)
-                df['kbps'] = df.iloc[:, 0:3].sum(axis=1).round(2) 
-                df['resolution'] = resolution
+                for i, video_name in enumerate(os.listdir(video_dir)):
+                    if i not in range(vid_start, vid_end + 1):
+                        continue
+                    start = perf_counter()
+                    print(f'Run {video_name} for person {person} resolution {resolution} quantizer {quantizer}')
+                    save_dir = f'{save_prefix}_vpx/{person}/resolution{resolution}/quantizer{quantizer}/' + \
+                            f'{os.path.basename(video_name)}/run0/'
+                    dump_file = f'{save_dir}/sender.log'
+                    saved_video_file = f'{save_dir}/received.mp4'
+                    print(save_dir)
 
-                per_frame_metrics = np.load(f'{save_dir}/metrics.npy', allow_pickle='TRUE').item()
-                averages = get_average_metrics(list(per_frame_metrics.values()))
-                metrics = {'psnr': [], 'ssim': [], 'lpips': [], 'latency': []}
-                for i, k in enumerate(metrics.keys()):
-                        metrics[k].append(averages[i])
+                    stats = log_parser.gather_trace_statistics(dump_file, args.window)
+                    num_windows = len(stats['bitrates']['video'])
+                    streams = list(stats['bitrates'].keys())
+                    stats['bitrates']['time'] = np.arange(1, num_windows + 1)
+                    window = stats['window']
+                    
+                    df = pd.DataFrame.from_dict(stats['bitrates'])
+                    for s in streams:
+                        df[s] = (df[s] / 1000.0 / window).round(2)
+                    df['kbps'] = df.iloc[:, 0:3].sum(axis=1).round(2) 
+                    df['resolution'] = resolution
 
-                for m in metrics.keys():
-                    while len(metrics[m]) < df.shape[0]:
-                        metrics[m].append(0)
-                    df[m] = metrics[m]
-                
-                combined_df = pd.concat([df, combined_df], ignore_index=True)
+                    per_frame_metrics = np.load(f'{save_dir}/metrics.npy', allow_pickle='TRUE').item()
+                    averages = get_average_metrics(list(per_frame_metrics.values()))
+                    metrics = {'psnr': [], 'ssim': [], 'lpips': [], 'latency': []}
+                    for i, k in enumerate(metrics.keys()):
+                            metrics[k].append(averages[i])
 
-            mean_df = pd.DataFrame(combined_df.mean(axis=0).round(2).to_dict(), index=[df.index.values[-1]])
-            mean_df['resolution'] = resolution
-            mean_df['quantizer'] = quantizer
-            if first:
-                mean_df.to_csv(args.csv_name, header=True, index=False, mode="w")
-                first = False
-            else:
-                mean_df.to_csv(args.csv_name, header=False, index=False, mode="a+")
+                    for m in metrics.keys():
+                        while len(metrics[m]) < df.shape[0]:
+                            metrics[m].append(metrics[m][0])
+                        df[m] = metrics[m]
+                    
+                    combined_df = pd.concat([df, combined_df], ignore_index=True)
+                    end = perf_counter()
+                    print("aggregating one piece of data", end - start)
+
+
+                mean_df = pd.DataFrame(combined_df.mean(axis=0).round(2).to_dict(), index=[df.index.values[-1]])
+                mean_df['resolution'] = resolution
+                mean_df['quantizer'] = quantizer
+                if first:
+                    mean_df.to_csv(args.csv_name, header=True, index=False, mode="w")
+                    first = False
+                else:
+                    mean_df.to_csv(args.csv_name, header=False, index=False, mode="a+")
 
 
 run_experiments()
